feat(ls-remote)!: make ls-remote a plain listing, move picker to install
pvm ls-remote now prints the remote versions (with installed markers
and package tags) instead of opening an install picker, so scripts
and non-TTY callers can consume it. The interactive picker moved to
'pvm install' without arguments, which already delegated here; it now
bails with a usage hint when stdin is not a terminal.

BREAKING CHANGE: 'pvm ls-remote' no longer prompts to install a
selected version. Use 'pvm install' without arguments for the
interactive picker.
