+++ src/commands/ls_remote.rs
@@ -2,7 +2,6 @@ use crate::{fs, network};
 use anyhow::Result;
 use clap::Parser;
 use colored::Colorize;
-use dialoguer::{Select, theme::ColorfulTheme};
 
 /// List all remote available PHP versions from static-php-cli
 #[derive(Parser, Debug)]
@@ -27,88 +26,20 @@ impl LsRemote {
 
         let installed = fs::list_installed_versions().unwrap_or_default();
 
-        let mut display_items = Vec::new();
-        let mut target_versions = Vec::new(); // Parallel array tying display index to actual version string
-
-        // Build "Quick Select" aliases
-        let mut minors = std::collections::BTreeMap::new();
-        let mut highest_overall = None;
-
-        for (v, _) in &versions_info {
-            highest_overall = Some(v.clone());
-            let parts: Vec<&str> = v.split('.').collect();
-            if parts.len() >= 2 {
-                let minor = format!("{}.{}", parts[0], parts[1]);
-                minors.insert(minor, v.clone()); // BTreeMap iterates ascending, keeps latest
-            }
-        }
-
-        if let Some(highest) = highest_overall {
-            display_items.push(format!("latest ({})", highest).bold().to_string());
-            target_versions.push(highest);
-        }
-
-        // Add them in reverse order (newest minor first) for the quick select
-        for (minor, highest_patch) in minors.iter().rev() {
-            display_items.push(
-                format!("{} ({})", minor, highest_patch)
-                    .bold()
-                    .cyan()
-                    .to_string(),
-            );
-            target_versions.push(highest_patch.clone());
-        }
-
-        display_items.push("---".dimmed().to_string());
-        target_versions.push("".to_string()); // Unselectable divider
-
-        // Build the rest of the flat list
-        for (v, pkgs) in versions_info.iter().rev() {
+        // Ascending order: the newest versions print last, right above the prompt.
+        for (v, pkgs) in &versions_info {
             let pkgs_str = pkgs.join(", ");
             if installed.contains(v) {
-                display_items.push(format!(
+                println!(
                     "{} {} {} [{}]",
                     "✓".green(),
                     v,
                     "(installed)".dimmed(),
                     pkgs_str.cyan()
-                ));
+                );
             } else {
-                display_items.push(format!("  {} [{}]", v, pkgs_str.cyan()));
+                println!("  {} [{}]", v, pkgs_str.cyan());
             }
-            target_versions.push(v.clone());
-        }
-
-        let selection = Select::with_theme(&ColorfulTheme::default())
-            .with_prompt("Select a PHP version to install")
-            .default(0)
-            .items(&display_items)
-            .interact_opt()?;
-
-        let selected = match selection {
-            Some(idx) => {
-                let target = &target_versions[idx];
-                if target.is_empty() {
-                    // They clicked the divider
-                    println!("{} Invalid selection.", "✗".red());
-                    return Ok(());
-                }
-                target
-            }
-            None => {
-                println!("{} Operation cancelled.", "✗".red());
-                return Ok(());
-            }
-        };
-
-        if !installed.contains(selected) {
-            crate::commands::install::execute_install_with(selected, true).await?;
-        } else {
-            println!(
-                "{} PHP {} is already installed.",
-                "✓".green(),
-                selected.bold()
-            );
         }
 
         Ok(())
